safari-534.53-branch/LayoutTests/canvas/philip/tests/2d.state.saverestore.font.html

<!DOCTYPE html>
<title>Canvas test: 2d.state.saverestore.font</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

// Test that restore() undoes any modifications
var old = ctx.font;
ctx.save();
ctx.font = "25px serif";
ctx.restore();
_assertSame(ctx.font, old, "ctx.font", "old");

// Also test that save() doesn't modify the values
ctx.font = "25px serif";
old = ctx.font;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against "25px serif"
ctx.save();
_assertSame(ctx.font, old, "ctx.font", "old");
ctx.restore();


});
</script>
