LayoutTests/canvas/philip/tests/2d.state.saverestore.font-expected.txt

< [index] >
2d.state.saverestore.font
save()/restore() works for font
References: 2d.state.font
Actual output:
Passed
